requirements.txt
    fastapi[all]
    uvicorn
    sqlalchemy
    mysqlclient
    passlib[bcrypt]
    python-jose[cryptography]
    python-dotenv
    